Refactored eslint-grunt.
With this refactoring eslint-grunt now respects the reporting level of the ESLint config JSON. If a rule is set to 2, the task will fail, 1 the task will warn, but succeed, 0 the rule is off.

No longer relying on console output to determine the success or failure of the task, Success/failure is based on CLI return value.

The task now supports custom formatters as well as the ESLint core formatters.